Move functions converting bytes to str to new mod
